extract to lib.svc*
