+++ scripts/md_to_html_text.py
@@ -6,11 +6,9 @@ Converts Markdown text to HTML and outputs to stdout
 
 import argparse
 import sys
+from pathlib import Path
 
-from pypandoc import convert_text
-
-
-from utils.utils import get_md_text
+from lib.svc_md_to_html_text import convert_md_to_html_text
 
 
 def main():
@@ -33,19 +31,12 @@ def main():
     except FileNotFoundError:
         md_text = args.input
 
-    # Process Markdown text
-    try:
-        md_text = get_md_text(md_text)
-    except ValueError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
     # Convert to HTML
     try:
-        html_str = convert_text(md_text, format="markdown", to="html").decode("utf-8")
+        html_str = convert_md_to_html_text(md_text)
         print(html_str)
     except Exception as e:
-        print(f"Error: Failed to convert to HTML - {e}", file=sys.stderr)
+        print(f"Error: {e}", file=sys.stderr)
         sys.exit(1)
