+++ scripts/md_to_xml.py
@@ -8,11 +8,7 @@ import argparse
 import sys
 from pathlib import Path
 
-import markdown
-from lxml import html, etree
-
-
-from utils.utils import get_md_text
+from lib.svc_md_to_xml import convert_md_to_xml
 
 
 def main():
@@ -43,30 +39,13 @@ def main():
     else:
         md_text = args.input
 
-    # Process Markdown text
-    try:
-        md_text = get_md_text(md_text, is_strip_wrapper=args.strip_wrapper)
-    except ValueError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
     # Convert to XML
     output_path = Path(args.output)
     try:
-        html_str = markdown.markdown(text=md_text, extensions=["extra", "toc"])
-        xml_element = html.fromstring(html_str)
-        result_file_bytes = etree.tostring(
-            element_or_tree=xml_element,
-            xml_declaration=True,
-            pretty_print=True,
-            encoding="UTF-8"
-        )
-
-        output_path.write_bytes(result_file_bytes)
-        print(f"Successfully converted to {output_path}")
-
+        output_file = convert_md_to_xml(md_text, output_path, args.strip_wrapper)
+        print(f"Successfully converted to {output_file}")
     except Exception as e:
-        print(f"Error: Failed to convert to XML - {e}", file=sys.stderr)
+        print(f"Error: {e}", file=sys.stderr)
         sys.exit(1)
