+++ tools/md_to_codeblock/md_to_codeblock.py
@@ -8,10 +8,10 @@ from dify_plugin import Tool
 from dify_plugin.entities.tool import ToolInvokeMessage
 
 from tools.md_to_codeblock.codeblock import CodeBlock
-from tools.utils.file_utils import get_meta_data
-from tools.utils.logger_utils import get_logger
-from tools.utils.mimetype_utils import MimeType
-from tools.utils.param_utils import get_md_text, get_param_value
+from scripts.utils.file_utils import get_meta_data
+from scripts.utils.logger_utils import get_logger
+from scripts.utils.mimetype_utils import MimeType
+from scripts.utils.param_utils import get_md_text, get_param_value
 
 
 class MarkdownToCodeblockTool(Tool):
