+++ tools/md_to_csv/md_to_csv.py
@@ -3,11 +3,11 @@ from typing import Generator, Optional
 from dify_plugin import Tool
 from dify_plugin.entities.tool import ToolInvokeMessage
 
-from tools.utils.file_utils import get_meta_data
-from tools.utils.logger_utils import get_logger
-from tools.utils.mimetype_utils import MimeType
-from tools.utils.param_utils import get_md_text
-from tools.utils.table_utils import TableParser
+from scripts.utils.file_utils import get_meta_data
+from scripts.utils.logger_utils import get_logger
+from scripts.utils.mimetype_utils import MimeType
+from scripts.utils.param_utils import get_md_text
+from scripts.utils.table_utils import TableParser
 
 
 from scripts.lib.svc_md_to_csv import get_csv_output_encoding
