+++ tools/md_to_html/md_to_html.py
@@ -3,14 +3,13 @@ from typing import Generator
 from dify_plugin import Tool
 from dify_plugin.entities.tool import ToolInvokeMessage
 
-from tools.utils.file_utils import get_meta_data
-from tools.utils.logger_utils import get_logger
-from tools.utils.mimetype_utils import MimeType
-from tools.utils.param_utils import get_md_text
+from scripts.utils.file_utils import get_meta_data
+from scripts.utils.logger_utils import get_logger
+from scripts.utils.mimetype_utils import MimeType
+from scripts.utils.param_utils import get_md_text
 
 
 from scripts.lib.svc_md_to_html import convert_md_to_html
-from scripts.utils.utils import get_md_text as get_md_text_base
 
 
 class MarkdownToHtmlTool(Tool):
