+++ tools/md_to_png/md_to_png.py
@@ -9,11 +9,11 @@ from PIL import Image
 from dify_plugin import Tool
 from dify_plugin.entities.tool import ToolInvokeMessage
 
-from tools.utils.file_utils import get_meta_data
-from tools.utils.logger_utils import get_logger
-from tools.utils.md_utils import MarkdownUtils
-from tools.utils.mimetype_utils import MimeType
-from tools.utils.param_utils import get_md_text, get_param_value
+from scripts.utils.file_utils import get_meta_data
+from scripts.utils.logger_utils import get_logger
+from scripts.utils.md_utils import MarkdownUtils
+from scripts.utils.mimetype_utils import MimeType
+from scripts.utils.param_utils import get_md_text, get_param_value
 
 
 class MarkdownToPngTool(Tool):
